make the branches and tags as input fields

+++ backend/releases/api.py
@@ -28,9 +28,6 @@ headers = {
 
 
 class SimpleConstantSchema(ModelSchema):
-    tags: list[str]
-    branches: list[str]
-
     class Config:
         model = Constant
         include = ("repo", "service", "name")
@@ -44,48 +41,6 @@ class ConstantResponse(Schema):
 @router.post("/constant", response=ConstantResponse)
 def post_constant(request, service_options: list[str]):
     constants = Constant.objects.filter(service__in=service_options).all()
-
-    for item in list(constants):
-        tags_response = requests.get(
-            f"https://api.github.com/repos/{item.repo}/tags?per_page=100",
-            headers=headers,
-        )
-
-        item.tags = list(map((lambda item: item["name"]), tags_response.json()))
-
-        if "Link" in tags_response.headers:
-            for index in range(
-                2,
-                int(tags_response.headers["Link"].split("&page=")[2].split(">")[0]) + 1,
-            ):
-                tags_response = requests.get(
-                    f"https://api.github.com/repos/{item.repo}/tags?per_page=100&page={index}",
-                    headers=headers,
-                )
-                item.tags = item.tags + list(
-                    map((lambda item: item["name"]), tags_response.json())
-                )
-
-        branches_response = requests.get(
-            f"https://api.github.com/repos/{item.repo}/branches?per_page=100",
-            headers=headers,
-        )
-
-        item.branches = list(map((lambda item: item["name"]), branches_response.json()))
-
-        if "Link" in branches_response.headers:
-            for index in range(
-                2,
-                int(branches_response.headers["Link"].split("&page=")[2].split(">")[0])
-                + 1,
-            ):
-                branches_response = requests.get(
-                    f"https://api.github.com/repos/{item.repo}/branches?per_page=100&page={index}",
-                    headers=headers,
-                )
-                item.branches = item.branches + list(
-                    map((lambda item: item["name"]), branches_response.json())
-                )
 
     return {
         "ok": True,
@@ -139,48 +94,6 @@ class ConstantUserResponse(Schema):
 def get_constant_and_users(request):
     constants = Constant.objects.all()
 
-    for item in list(constants):
-        tags_response = requests.get(
-            f"https://api.github.com/repos/{item.repo}/tags?per_page=100",
-            headers=headers,
-        )
-
-        item.tags = list(map((lambda item: item["name"]), tags_response.json()))
-
-        if "Link" in tags_response.headers:
-            for index in range(
-                2,
-                int(tags_response.headers["Link"].split("&page=")[2].split(">")[0]) + 1,
-            ):
-                tags_response = requests.get(
-                    f"https://api.github.com/repos/{item.repo}/tags?per_page=100&page={index}",
-                    headers=headers,
-                )
-                item.tags = item.tags + list(
-                    map((lambda item: item["name"]), tags_response.json())
-                )
-
-        branches_response = requests.get(
-            f"https://api.github.com/repos/{item.repo}/branches?per_page=100",
-            headers=headers,
-        )
-
-        item.branches = list(map((lambda item: item["name"]), branches_response.json()))
-
-        if "Link" in branches_response.headers:
-            for index in range(
-                2,
-                int(branches_response.headers["Link"].split("&page=")[2].split(">")[0])
-                + 1,
-            ):
-                branches_response = requests.get(
-                    f"https://api.github.com/repos/{item.repo}/branches?per_page=100&page={index}",
-                    headers=headers,
-                )
-                item.branches = item.branches + list(
-                    map((lambda item: item["name"]), branches_response.json())
-                )
-
     users = Account.objects.filter(~Q(first_name=""))
     release_list = Release.objects.all()
 
@@ -212,39 +125,57 @@ class CreateReleaseRequest(Schema):
 @router.post("/create", response=AckResponse)
 def create_release(request, form: CreateReleaseRequest):
     user = request.auth
+    false_branches = []
     try:
         for role in user.roles.all():
             if role.role == Roles.Role.ReleaseAdmin:
                 with transaction.atomic():
-                    release = Release.objects.create(
-                        name=form.release.name,
-                        created_by=user,
-                        updated_by=user,
-                        start_window=form.release.start_window,
-                        end_window=form.release.end_window,
-                    )
                     for item in form.release.items:
-                        ReleaseItem.objects.create(
-                            repo=item.repo,
-                            service=item.service,
-                            release_branch=item.release_branch,
-                            hotfix_branch=item.hotfix_branch,
-                            tag=item.tag,
-                            special_notes=item.special_notes,
-                            devops_notes=item.devops_notes,
-                            release=release,
+                        if not item.release_branch:
+                            continue
+                        git_response = requests.get(
+                            f"https://api.github.com/repos/{item.repo}/branches/{item.release_branch}",
+                            headers=headers,
                         )
-                    for item in form.approvers:
-                        Approver.objects.create(group=item, release=release)
-                    Approver.objects.create(group=2, release=release)
-                    Approver.objects.create(group=4, release=release)
-                    for item in form.targets:
-                        Target.objects.create(target=item, release=release)
+                        if git_response.status_code == 404:
+                            false_branches.append(item.repo)
+                    if len(false_branches):
+                        return {
+                            "ok": False,
+                            "error": {
+                                "detail": ", ".join(false_branches),
+                                "reason": "branch_not_found",
+                            },
+                        }
+                    else:
+                        release = Release.objects.create(
+                            name=form.release.name,
+                            created_by=user,
+                            updated_by=user,
+                            start_window=form.release.start_window,
+                            end_window=form.release.end_window,
+                        )
+                        for item in form.release.items:
+                            ReleaseItem.objects.create(
+                                repo=item.repo,
+                                service=item.service,
+                                release_branch=item.release_branch,
+                                hotfix_branch=item.hotfix_branch,
+                                tag=item.tag,
+                                special_notes=item.special_notes,
+                                devops_notes=item.devops_notes,
+                                release=release,
+                            )
+                        for item in form.approvers:
+                            Approver.objects.create(group=item, release=release)
+                        Approver.objects.create(group=2, release=release)
+                        Approver.objects.create(group=4, release=release)
+                        for item in form.targets:
+                            Target.objects.create(target=item, release=release)
 
                 return {"ok": True}
     except Exception as e:
-        logger.error(e)
-        {
+        return {
             "ok": False,
             "error": {
                 "reason": "internal_server_error",
@@ -283,6 +214,7 @@ class UpdateReleaseRequest(Schema):
 @router.post("/update", response=AckResponse)
 def update_release(request, form: UpdateReleaseRequest):
     user = request.auth
+    false_branches = []
     release = Release.objects.get(uuid=form.uuid)
 
     for approver in list(release.approvers.all()):
@@ -292,31 +224,49 @@ def update_release(request, form: UpdateReleaseRequest):
             continue
         else:
             with transaction.atomic():
-                release.updated_by = user
-                release.start_window = form.release.start_window
-                release.end_window = form.release.end_window
-                release.deployment_status = form.release.deployment_status
-                release.deployment_comment = form.release.deployment_comment
-                release.save()
                 for item in form.release.items:
-                    release_item = ReleaseItem.objects.get(
-                        release=release,
-                        repo=item.repo,
-                        service=item.service,
+                    if not item.release_branch:
+                        continue
+                    git_response = requests.get(
+                        f"https://api.github.com/repos/{item.repo}/branches/{item.release_branch}",
+                        headers=headers,
                     )
-                    release_item.release_branch = item.release_branch
-                    release_item.hotfix_branch = item.hotfix_branch
-                    release_item.tag = item.tag
-                    release_item.special_notes = item.special_notes
-                    if Roles.Role.DevOps in [
-                        role.role for role in list(user.roles.all())
-                    ]:
-                        release_item.devops_notes = item.devops_notes
-                    release_item.save()
-                targets = Target.objects.filter(release=release)
-                targets.delete()
-                for item in form.targets:
-                    Target.objects.create(target=item, release=release)
+                    if git_response.status_code == 404:
+                        false_branches.append(item.repo)
+                if len(false_branches):
+                    return {
+                        "ok": False,
+                        "error": {
+                            "detail": ", ".join(false_branches),
+                            "reason": "branch_not_found",
+                        },
+                    }
+                else:
+                    release.updated_by = user
+                    release.start_window = form.release.start_window
+                    release.end_window = form.release.end_window
+                    release.deployment_status = form.release.deployment_status
+                    release.deployment_comment = form.release.deployment_comment
+                    release.save()
+                    for item in form.release.items:
+                        release_item = ReleaseItem.objects.get(
+                            release=release,
+                            repo=item.repo,
+                            service=item.service,
+                        )
+                        release_item.release_branch = item.release_branch
+                        release_item.hotfix_branch = item.hotfix_branch
+                        release_item.tag = item.tag
+                        release_item.special_notes = item.special_notes
+                        if Roles.Role.DevOps in [
+                            role.role for role in list(user.roles.all())
+                        ]:
+                            release_item.devops_notes = item.devops_notes
+                        release_item.save()
+                    targets = Target.objects.filter(release=release)
+                    targets.delete()
+                    for item in form.targets:
+                        Target.objects.create(target=item, release=release)
 
             return {"ok": True}
     return {
@@ -419,48 +369,6 @@ def get_release_with_uuid(request, uuid: uuid.UUID):
         service_options[item.service] = True
 
     constants = Constant.objects.filter(service__in=service_options.keys()).all()
-
-    for item in list(constants):
-        tags_response = requests.get(
-            f"https://api.github.com/repos/{item.repo}/tags?per_page=100",
-            headers=headers,
-        )
-
-        item.tags = list(map((lambda item: item["name"]), tags_response.json()))
-
-        if "Link" in tags_response.headers:
-            for index in range(
-                2,
-                int(tags_response.headers["Link"].split("&page=")[2].split(">")[0]) + 1,
-            ):
-                tags_response = requests.get(
-                    f"https://api.github.com/repos/{item.repo}/tags?per_page=100&page={index}",
-                    headers=headers,
-                )
-                item.tags = item.tags + list(
-                    map((lambda item: item["name"]), tags_response.json())
-                )
-
-        branches_response = requests.get(
-            f"https://api.github.com/repos/{item.repo}/branches?per_page=100",
-            headers=headers,
-        )
-
-        item.branches = list(map((lambda item: item["name"]), branches_response.json()))
-
-        if "Link" in branches_response.headers:
-            for index in range(
-                2,
-                int(branches_response.headers["Link"].split("&page=")[2].split(">")[0])
-                + 1,
-            ):
-                branches_response = requests.get(
-                    f"https://api.github.com/repos/{item.repo}/branches?per_page=100&page={index}",
-                    headers=headers,
-                )
-                item.branches = item.branches + list(
-                    map((lambda item: item["name"]), branches_response.json())
-                )
 
     return {
         "ok": True,
